geneEd/requirements.txt
certifi==2020.6.20
click==7.1.2
entrezpy==2.1.3
Flask==1.1.2
Flask-WTF==0.15.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
mysql-connector-python==8.0.18
protobuf==3.6.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn==20.0.4
beautifulsoup4==4.9.3